requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
boto3==1.35.20
httpx==0.27.2
sqlmodel==0.0.22
sqlalchemy==2.0.32
python-dotenv==1.0.1
httpx>=0.27